Fix cdb broadcast panic
The problem was double setting of the value on the reservation station
due to reuse of physical registers.

The solution is to add a check if the value is already set; but once it
is set, its physical register is meaningless (invalid) and this will
protect against setting the value again.